Refactor SPA to Context+Worker+Router architecture
- Migrated heavy geo logic to `src/workers/geo.worker.js`.
- Implemented `GlobalContext.jsx` for Auth, User Data, and Geo state.
- Implemented `MainLayout.jsx` and `main.jsx` using `react-router-dom`.
- Extracted Views to `src/pages/` (Records, Stats, Login, TripEditor).
- Refactored `RailRound.jsx` into `src/components/MapComponent.jsx`.
- Cleaned up obsolete files.
