+++ src/components/Tutorial.jsx
@@ -1,11 +1,13 @@
 import React, { useState, useEffect, useRef, useLayoutEffect, useCallback } from 'react';
 import { createPortal } from 'react-dom';
 import { X, ChevronRight, CheckCircle2, ArrowRight } from 'lucide-react';
+import { useLocation, useNavigate } from 'react-router-dom';
+import { useAuth, useGeo, useUserData } from '../globalContext';
 
 const STEPS = [
     {
         id: 'welcome',
-        target: '#header-title', // Center modal
+        target: null,
         title: "欢迎来到 RailLOOP",
         content: "RailLOOP 是一个个人向旅铁手账, 旨在帮助你追踪和管理你的铁路旅程, 直观可感地展示旅行足迹.",
         position: 'center',
@@ -13,12 +15,12 @@ const STEPS = [
     },
     {
         id: 'tab-records',
-        target: '#tab-btn-records',
+        target: 'nav a[href="/records"]', // Selector for Records Tab
         title: "行程记录",
         content: "这个标签页是你的旅程控制中心. 你可以在这里查看、添加和管理所有的铁路旅行记录, 或者添加到收藏",
         position: 'top',
         action: 'switch-tab',
-        tab: 'records'
+        path: '/records'
     },
     {
         id: 'add-trip',
@@ -40,7 +42,7 @@ const STEPS = [
     },
     {
         id: 'auto-planning-view',
-        target: '#auto-planning-form',
+        target: '#auto-planning-form', // This ID must exist in TripEditorPage
         title: "自动规划",
         content: "在这里选择起点和终点，系统会自动为你规划一条基于新干线优先的路线。完成后，我们继续。",
         position: 'bottom',
@@ -48,7 +50,7 @@ const STEPS = [
     },
     {
         id: 'close-editor',
-        target: '#btn-close-editor',
+        target: '#btn-close-editor', // Needs ID in TripEditorPage
         title: "关闭编辑器",
         content: "先关掉它. 继续参观飞船的其他部分，别让编辑器挡住了视线. ",
         position: 'bottom',
@@ -65,12 +67,12 @@ const STEPS = [
     },
     {
         id: 'tab-map',
-        target: '#tab-btn-map',
+        target: 'nav a[href="/map"]',
         title: "地图模式",
         content: "切换并查看铁路网络. 灰色的是未乘区段, 而已乘线路将以对应的颜色高亮显示. 放大到 10x 可查看 openrailwaymap 提供的配线和站台详情.",
         position: 'top',
         action: 'switch-tab',
-        tab: 'map'
+        path: '/map'
     },
     {
         id: 'map-pins',
@@ -100,13 +102,12 @@ const STEPS = [
     },
     {
         id: 'tab-stats',
-        target: '#tab-btn-stats',
+        target: 'nav a[href="/stats"]',
         title: "Aftermath",
         content: "点击这里查看统计。看看你在这个星球的铁轨上究竟烧掉了多少钱，以及产生了多少碳排放(已加入todo).",
         position: 'top',
-        action: 'wait-click-tab', // Specifically asked to guide click
-        tab: 'stats',
-        check: ({ activeTab }) => activeTab === 'stats'
+        action: 'switch-tab',
+        path: '/stats'
     },
     {
         id: 'stats-content',
@@ -126,7 +127,7 @@ const STEPS = [
     },
     {
          id: 'login-guide',
-         target: '#login-readme-container',
+         target: '#login-readme-container', // Login Page needs this ID?
          title: "阅读协议",
          content: "在开始之前，请务必阅读这份实则是免责声明的用户指南. 如果你们当中有宇宙人、未来人、超能力者, 请忽略其中的物理限制条款.",
          position: 'left',
@@ -134,38 +135,34 @@ const STEPS = [
     }
 ];
 
-const Tutorial = ({
-    activeTab,
-    setActiveTab,
-    isTripEditing,
-    setIsTripEditing,
-    isLoginOpen,
-    setIsLoginOpen,
-    user,
-    pinMode,
-    editorMode
-}) => {
-    const [step, setStep] = useState(-1); // -1: Loading/Check, 0+: Steps
+const Tutorial = () => {
+    const { user } = useAuth();
+    const { pinMode } = useGeo();
+    const { editorMode } = useUserData();
+    const location = useLocation();
+    const navigate = useNavigate();
+
+    const [step, setStep] = useState(-1);
     const [rect, setRect] = useState(null);
     const [isVisible, setIsVisible] = useState(false);
     const [tooltipStyle, setTooltipStyle] = useState({});
-
-    // Ref for the tooltip card to measure strict dimensions
     const tooltipRef = useRef(null);
 
+    // Derived States
+    const isTripEditing = location.pathname.startsWith('/trips/');
+    const isLoginOpen = location.pathname === '/login';
+
     const handleNext = useCallback(() => {
-        if (step === STEPS.length - 2) { // Finish step
-            // Open Login Modal
-            setIsLoginOpen(true);
+        if (step === STEPS.length - 2) {
+            navigate('/login');
             setTimeout(() => setStep(s => s + 1), 200);
         } else if (step >= STEPS.length - 1) {
-            // End
             setStep(-2);
             setIsVisible(false);
         } else {
             setStep(s => s + 1);
         }
-    }, [step, setIsLoginOpen, setStep, setIsVisible]);
+    }, [step, navigate]);
 
     const handleSkip = useCallback((dontShowAgain) => {
         if (dontShowAgain) {
@@ -173,101 +170,65 @@ const Tutorial = ({
         }
         setStep(-2);
         setIsVisible(false);
-    }, [setStep, setIsVisible]);
+    }, []);
 
-    // Keyboard navigation for tutorial
     useEffect(() => {
-        if (!isVisible || step < 0 || step >= STEPS.length) return;
-
         const handleKeyDown = (e) => {
-            if (e.key === 'Escape') {
-                handleSkip(false);
-            } else if (e.key === 'Enter') {
+            if (!isVisible) return;
+            if (e.key === 'Escape') handleSkip(false);
+            else if (e.key === 'Enter') {
                 const currentStepConfig = STEPS[step];
-                if (currentStepConfig.action !== 'wait-interaction' && currentStepConfig.action !== 'wait-click-tab') {
+                if (currentStepConfig && currentStepConfig.action !== 'wait-interaction') {
                     handleNext();
                 }
             }
         };
-
         document.addEventListener('keydown', handleKeyDown);
         return () => document.removeEventListener('keydown', handleKeyDown);
     }, [isVisible, step, handleNext, handleSkip]);
 
-    // Initialization check
     useEffect(() => {
         const skipped = localStorage.getItem('rail_tutorial_skipped');
-
         if (skipped === 'true' || user) {
-            setStep(-2); // Skipped
+            setStep(-2);
             return;
         }
         setStep(0);
         setIsVisible(true);
     }, [user]);
 
-    // Step Transition Logic & Rect Calculation
     useEffect(() => {
         if (step < 0 || step >= STEPS.length) return;
-
         const currentStep = STEPS[step];
 
-        // 1. Target Resolution
         const updateRect = () => {
-            if (!currentStep.target ) {
+            if (!currentStep.target) {
                 setRect(null);
                 setTooltipStyle({
-                    position: 'fixed',
-                    top: '50%',
-                    left: '50%',
-                    transform: 'translate(-50%, -50%)',
-                    opacity: 1,
-                    transition: 'all 0.5s cubic-bezier(0.25, 1, 0.5, 1)'
+                    position: 'fixed', top: '50%', left: '50%', transform: 'translate(-50%, -50%)', opacity: 1, transition: 'all 0.5s'
                 });
                 return;
             }
-            if (currentStep.target === '#header-title') {
-                setTooltipStyle({
-                    position: 'fixed',
-                    top: '50%',
-                    left: '50%',
-                    transform: 'translate(-50%, -50%)',
-                    opacity: 1,
-                    transition: 'all 0.5s cubic-bezier(0.25, 1, 0.5, 1)'
-                });
-            }    
             const el = document.querySelector(currentStep.target);
             if (el) {
                 const r = el.getBoundingClientRect();
-
-                // Add padding to highlight box
-                const highlight = {
-                    top: r.top - 5,
-                    left: r.left - 5,
-                    width: r.width + 10,
-                    height: r.height + 10
-                };
-                setRect(highlight);
+                setRect({ top: r.top - 5, left: r.left - 5, width: r.width + 10, height: r.height + 10 });
             } else {
                 setRect(null);
             }
         };
 
-        // Initial update
         updateRect();
-        // Poll for rect changes (animations)
         const interval = setInterval(updateRect, 100);
 
-        if (currentStep.action === 'switch-tab' && currentStep.tab && activeTab !== currentStep.tab) {
-             setActiveTab(currentStep.tab);
+        if (currentStep.action === 'switch-tab' && currentStep.path && location.pathname !== currentStep.path) {
+             navigate(currentStep.path);
         }
 
-        // 3. Wait Conditions
         let checkInterval;
         if (currentStep.check) {
             checkInterval = setInterval(() => {
-                // Pass component state to the check function
-                if (currentStep.check({ activeTab, isTripEditing, isLoginOpen, pinMode, editorMode })) {
+                if (currentStep.check({ isTripEditing, isLoginOpen, pinMode, editorMode })) {
                     handleNext();
                 }
             }, 200);
@@ -277,194 +238,93 @@ const Tutorial = ({
             clearInterval(interval);
             if (checkInterval) clearInterval(checkInterval);
         };
-    }, [step, activeTab, isTripEditing, isLoginOpen, pinMode, editorMode]);
+    }, [step, isTripEditing, isLoginOpen, pinMode, editorMode, location.pathname, navigate, handleNext]);
 
-    // Strict Positioning Logic
+
     const calculatePosition = () => {
         if (step < 0 || step >= STEPS.length || !rect || !tooltipRef.current) return;
 
         const currentStep = STEPS[step];
         const PADDING = 20;
-
-        // Measure ACTUAL tooltip dimensions
         const CARD_W = tooltipRef.current.offsetWidth || 384;
         const CARD_H = tooltipRef.current.offsetHeight || 250;
-
-        // Viewport Dimensions
-        const winH = window.visualViewport ? window.visualViewport.height : window.innerHeight;
-        const winW = window.visualViewport ? window.visualViewport.width : window.innerWidth;
-        const offsetTop = window.visualViewport ? window.visualViewport.offsetTop : 0;
-
-        // Adjust for viewport scrolling/offset if using visualViewport
-        // Actually for fixed position elements, visualViewport logic is complex.
-        // If we use simple innerHeight, it might be safer unless keyboard is up.
-        // But for FAB (bottom) clipping, visualViewport height is the visible part.
-
-        // Center of the target element
+        const winH = window.innerHeight;
+        const winW = window.innerWidth;
         const targetCenterX = rect.left + rect.width / 2;
         const targetCenterY = rect.top + rect.height / 2;
 
-        let top = 0;
-        let left = 0;
-        let pos = currentStep.position;
-
-        // Helper to check collision
-        const checkBounds = (t, l) => {
-            return (t >= PADDING && l >= PADDING && (t + CARD_H) <= (winH - PADDING) && (l + CARD_W) <= (winW - PADDING));
-        };
-
-        // 1. Determine Initial Position & Flip if needed
         const getCoords = (p) => {
             let t, l;
-            if (p === 'bottom') {
-                t = rect.top + rect.height + 20;
-                l = targetCenterX - (CARD_W / 2);
-            } else if (p === 'top') {
-                t = rect.top - CARD_H - 20;
-                l = targetCenterX - (CARD_W / 2);
-            } else if (p === 'right') {
-                t = targetCenterY - (CARD_H / 2);
-                l = rect.left + rect.width + 20;
-            } else if (p === 'left') {
-                t = targetCenterY - (CARD_H / 2);
-                l = rect.left - CARD_W - 20;
-            } else { // center
-                t = winH / 2 - CARD_H / 2;
-                l = winW / 2 - CARD_W / 2;
-            }
+            if (p === 'bottom') { t = rect.top + rect.height + 20; l = targetCenterX - (CARD_W / 2); }
+            else if (p === 'top') { t = rect.top - CARD_H - 20; l = targetCenterX - (CARD_W / 2); }
+            else if (p === 'right') { t = targetCenterY - (CARD_H / 2); l = rect.left + rect.width + 20; }
+            else if (p === 'left') { t = targetCenterY - (CARD_H / 2); l = rect.left - CARD_W - 20; }
+            else { t = winH / 2 - CARD_H / 2; l = winW / 2 - CARD_W / 2; }
             return { t, l };
         };
 
+        let pos = currentStep.position;
         let coords = getCoords(pos);
 
-        // Smart Flip
+        // Simple Flip Check
+        const checkBounds = (t, l) => (t >= PADDING && l >= PADDING && (t + CARD_H) <= (winH - PADDING) && (l + CARD_W) <= (winW - PADDING));
+
         if (!checkBounds(coords.t, coords.l) && pos !== 'center') {
             const opposites = { 'top': 'bottom', 'bottom': 'top', 'left': 'right', 'right': 'left' };
             const altPos = opposites[pos];
             if (altPos) {
                 const altCoords = getCoords(altPos);
-                // If alternative is better (or valid), take it
-                // Simple heuristic: check if alt fits vertically for top/bottom swap
-                if (pos === 'top' || pos === 'bottom') {
-                    if (altCoords.t >= PADDING && (altCoords.t + CARD_H) <= (winH - PADDING)) {
-                        coords = altCoords;
-                        pos = altPos;
-                    }
-                }
-                // Check if alt fits horizontally for left/right swap
-                if (pos === 'left' || pos === 'right') {
-                    if (altCoords.l >= PADDING && (altCoords.l + CARD_W) <= (winW - PADDING)) {
-                        coords = altCoords;
-                        pos = altPos;
-                    }
+                if (checkBounds(altCoords.t, altCoords.l)) {
+                    coords = altCoords;
+                    pos = altPos;
                 }
             }
         }
 
-        // 2. Strict Clamping (The "Aggressive" Part)
-        // Ensure card never goes off screen, even if it de-centers from target
-        top = Math.max(PADDING, Math.min(coords.t, winH - CARD_H - PADDING));
-        left = Math.max(PADDING, Math.min(coords.l, winW - CARD_W - PADDING));
+        const top = Math.max(PADDING, Math.min(coords.t, winH - CARD_H - PADDING));
+        const left = Math.max(PADDING, Math.min(coords.l, winW - CARD_W - PADDING));
 
         setTooltipStyle({
-            position: 'fixed',
-            top: `${top}px`,
-            left: `${left}px`,
-            transform: 'none',
-            opacity: 1,
-            transition: 'all 0.4s cubic-bezier(0.25, 1, 0.5, 1)' // Smooth transitions
+            position: 'fixed', top: `${top}px`, left: `${left}px`, transform: 'none', opacity: 1, transition: 'all 0.4s'
         });
     };
 
-    useLayoutEffect(() => {
-        calculatePosition();
-    }, [rect, step]);
-
-    // ResizeObserver to handle content size changes
+    useLayoutEffect(() => { calculatePosition(); }, [rect, step]);
     useEffect(() => {
         if (!tooltipRef.current) return;
-        const observer = new ResizeObserver(() => {
-            calculatePosition();
-        });
+        const observer = new ResizeObserver(() => calculatePosition());
         observer.observe(tooltipRef.current);
         return () => observer.disconnect();
     }, [rect, step]);
 
     if (!isVisible || step < 0) return null;
-
     const currentStep = STEPS[step];
 
-    // Render
     return createPortal(
         <div className="fixed inset-0 z-[2000] overflow-hidden pointer-events-none">
-            {/* Mask / Spotlight */}
             {rect ? (
-                <div
-                    className="absolute transition-all duration-300 ease-out border-gray-900/80"
-                    style={{
-                        top: rect.top,
-                        left: rect.left,
-                        width: rect.width,
-                        height: rect.height,
-                        boxShadow: '0 0 0 9999px rgba(0, 0, 0, 0.75)',
-                        borderRadius: '8px',
-                        pointerEvents: 'none' // Let clicks pass through to target
-                    }}
+                <div className="absolute transition-all duration-300 ease-out border-gray-900/80"
+                    style={{ top: rect.top, left: rect.left, width: rect.width, height: rect.height, boxShadow: '0 0 0 9999px rgba(0, 0, 0, 0.75)', borderRadius: '8px', pointerEvents: 'none' }}
                 />
-            ) : (
-                // Full overlay if no target (Welcome step)
-                <div className="absolute inset-0 bg-black/75 pointer-events-auto" />
-            )}
+            ) : (<div className="absolute inset-0 bg-black/75 pointer-events-auto" />)}
 
-            {/* Tooltip / Card */}
-            <div
-                ref={tooltipRef}
-                className={`absolute pointer-events-auto flex flex-col gap-4 max-w-sm w-full p-6 bg-white rounded-2xl shadow-2xl`}
-                style={{
-                    ...tooltipStyle,
-                    maxWidth: 'min(24rem, calc(100vw - 40px))'
-                }}
-            >
+            <div ref={tooltipRef} className={`absolute pointer-events-auto flex flex-col gap-4 max-w-sm w-full p-6 bg-white rounded-2xl shadow-2xl`} style={{ ...tooltipStyle, maxWidth: 'min(24rem, calc(100vw - 40px))' }}>
                 <div>
-                    <div className="flex justify-between items-start mb-2">
-                        <h3 className="text-xl font-bold text-gray-800">{currentStep.title}</h3>
-                        <span className="text-xs font-bold text-gray-400 px-2 py-1 bg-gray-100 rounded-full">{step + 1} / {STEPS.length}</span>
-                    </div>
-                    <p className="text-gray-600 text-sm leading-relaxed mb-6">
-                        {currentStep.content}
-                    </p>
-
+                    <div className="flex justify-between items-start mb-2"><h3 className="text-xl font-bold text-gray-800">{currentStep.title}</h3><span className="text-xs font-bold text-gray-400 px-2 py-1 bg-gray-100 rounded-full">{step + 1} / {STEPS.length}</span></div>
+                    <p className="text-gray-600 text-sm leading-relaxed mb-6">{currentStep.content}</p>
                     <div className="flex items-center justify-between">
                          <div className="flex items-center gap-2">
-                            {step === 0 && (
-                                <label className="flex items-center gap-2 text-xs text-gray-500 cursor-pointer hover:text-gray-700">
-                                    <input type="checkbox" onChange={(e) => { if(e.target.checked) handleSkip(true); }} className="rounded border-gray-300"/>
-                                    不再显示
-                                </label>
-                            )}
-                            {step > 0 && (
-                                <button onClick={() => handleSkip(true)} className="text-xs text-gray-400 hover:text-gray-600 underline">跳过教程</button>
-                            )}
+                            {step === 0 && (<label className="flex items-center gap-2 text-xs text-gray-500 cursor-pointer hover:text-gray-700"><input type="checkbox" onChange={(e) => { if(e.target.checked) handleSkip(true); }} className="rounded border-gray-300"/> 不再显示</label>)}
+                            {step > 0 && (<button onClick={() => handleSkip(true)} className="text-xs text-gray-400 hover:text-gray-600 underline">跳过教程</button>)}
                          </div>
-
-                        {currentStep.action !== 'wait-interaction' && currentStep.action !== 'wait-click-tab' && (
-                             <button
-                                onClick={handleNext}
-                                className="bg-emerald-600 hover:bg-emerald-700 text-white px-5 py-2.5 rounded-xl font-bold text-sm flex items-center gap-2 shadow-lg shadow-emerald-200 transition-all transform active:scale-95"
-                             >
-                                {step === STEPS.length - 1 ? '结束' : '继续'} <ArrowRight size={16}/>
-                             </button>
-                        )}
-                        {(currentStep.action === 'wait-interaction' || currentStep.action === 'wait-click-tab') && (
-                            <div className="text-xs font-bold text-emerald-600 animate-pulse flex items-center gap-1">
-                                <ChevronRight size={14}/> 请按照指示操作
-                            </div>
+                        {currentStep.action !== 'wait-interaction' && currentStep.action !== 'wait-click-tab' ? (
+                             <button onClick={handleNext} className="bg-emerald-600 hover:bg-emerald-700 text-white px-5 py-2.5 rounded-xl font-bold text-sm flex items-center gap-2 shadow-lg shadow-emerald-200 transition-all transform active:scale-95">{step === STEPS.length - 1 ? '结束' : '继续'} <ArrowRight size={16}/></button>
+                        ) : (
+                            <div className="text-xs font-bold text-emerald-600 animate-pulse flex items-center gap-1"><ChevronRight size={14}/> 请按照指示操作</div>
                         )}
                     </div>
                 </div>
             </div>
-        </div>
-    , document.body);
+        </div>, document.body);
 };
-
 export default Tutorial;
